LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-waits-for-activate.https-expected.txt



FAIL Fetch events should wait for the activate event to complete. assert_unreached: unexpected rejection: assert_equals: frame should not be loaded expected (undefined) undefined but got (object) Element node <iframe class="test-iframe" src="resources/fetch-waits-fo... Reached unreachable code
